Changed Game's millisSince to return a float
This is to prevent the system breaking down over 1000 updates per second

+++ WarpDrive/basesystem/displaymanager.cpp
@@ -124,7 +124,7 @@ void DisplayManager::Update()
     lastCall = Game::instance()->now();
 }
 
-int64_t DisplayManager::Dt() const noexcept
+float DisplayManager::Dt() const noexcept
 {
     return dt;
 }
